NFS: Fix an Oopsable condition in __nfs_pageio_add_request()
Git-commit: ee21cd3aa8548e0cbc8c67a80b62113aedd2d101
References: bsn#19
Patch-mainline: v5.10.42

commit 56517ab958b7c11030e626250c00b9b1a24b41eb upstream.

Ensure that nfs_pageio_error_cleanup() resets the mirror array contents,
so that the structure reflects the fact that it is now empty.
Also change the test in nfs_pageio_do_add_request() to be more robust by
checking whether or not the list is empty rather than relying on the
value of pg_count.

Fixes: a7d42ddb3099 ("nfs: add mirroring support to pgio layer")
Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.42-0093-NFS-Fix-an-Oopsable-condition-in-__nfs_pageio_add_re.patch