mptcp: restrict values of 'enabled' sysctl
Git-commit: 744ee14054c8ca5ad0fe3ab9172709c17d8a240a
Message-Id: <744ee14054c8ca5ad0fe3ab9172709c17d8a240a.1637072791.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.14-rc1

To avoid confusions, it seems better to parse this sysctl parameter as a
boolean. We use it as a boolean, no need to parse an integer and bring
confusions if we see a value different from 0 and 1, especially with
this parameter name: enabled.

It seems fine to do this modification because the default value is 1
(enabled). Then the only other interesting value to set is 0 (disabled).
All other values would not have changed the default behaviour.

Suggested-by: Florian Westphal <fw@strlen.de>
Acked-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.14-rc1-mptcp-restrict-values-of-enabled-sysctl.patch