drm/amd/display: Avoid HDCP over-read and corruption
Git-commit: c5b518f4b98dbb2bc31b6a55e6aaa1e0e2948f2e
References: bsn#19
Patch-mainline: v5.10.51

[ Upstream commit 06888d571b513cbfc0b41949948def6cb81021b2 ]

Instead of reading the desired 5 bytes of the actual target field,
the code was reading 8. This could result in a corrupted value if the
trailing 3 bytes were non-zero, so instead use an appropriately sized
and zero-initialized bounce buffer, and read only 5 bytes before casting
to u64.

Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.51-0047-drm-amd-display-Avoid-HDCP-over-read-and-corruption.patch