xfs: detach inode dquots at the end of inactivation
From 3ea06d73e3c02ee2952a62bf92abc18f9c98aba1 Mon Sep 17 00:00:00 2001
Git-commit: 3ea06d73e3c02ee2952a62bf92abc18f9c98aba1
Patch-mainline: v5.14-rc1
References: bsn#12
Modified-by-SEL: No

Once we're done with inactivating an inode, we're finished updating
metadata for that inode.  This means that we can detach the dquots at
the end and not have to wait for reclaim to do it for us.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.mainline/v5.14-rc1-xfs-detach-inode-dquots-at-the-end-of-inactivation.patch