net/tls: Replace TLS_RX_SYNC_RUNNING with RCU
Git-commit: 874ece252ed269f5ac1f55167a3f2735ab0f249f
References: bsn#19
Patch-mainline: v5.10.43

[ Upstream commit 05fc8b6cbd4f979a6f25759c4a17dd5f657f7ecd ]

RCU synchronization is guaranteed to finish in finite time, unlike a
busy loop that polls a flag. This patch is a preparation for the bugfix
in the next patch, where the same synchronize_net() call will also be
used to sync with the TX datapath.

Signed-off-by: Maxim Mikityanskiy <maximmi@nvidia.com>
Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.43-0027-net-tls-Replace-TLS_RX_SYNC_RUNNING-with-RCU.patch