NFS: Fix a potential NULL dereference in nfs_get_client()
Git-commit: 279ad78a00f8b9c5ff24171a59297187a3bd44b7
References: bsn#19
Patch-mainline: v5.10.44

[ Upstream commit 09226e8303beeec10f2ff844d2e46d1371dc58e0 ]

None of the callers are expecting NULL returns from nfs_get_client() so
this code will lead to an Oops.  It's better to return an error
pointer.  I expect that this is dead code so hopefully no one is
affected.

Fixes: 31434f496abb ("nfs: check hostname in nfs_get_client")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.44-0118-NFS-Fix-a-potential-NULL-dereference-in-nfs_get_clie.patch