xfrm: Fix error reporting in xfrm_state_construct.
Git-commit: 57f7ed25bd1672a51c2d9cd15c9714c0dfc56062
References: bsn#19
Patch-mainline: v5.10.51

[ Upstream commit 6fd06963fa74197103cdbb4b494763127b3f2f34 ]

When memory allocation for XFRMA_ENCAP or XFRMA_COADDR fails,
the error will not be reported because the -ENOMEM assignment
to the err variable is overwritten before. Fix this by moving
these two in front of the function so that memory allocation
failures will be reported.

Reported-by: Tobias Brunner <tobias@strongswan.org>
Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.51-0077-xfrm-Fix-error-reporting-in-xfrm_state_construct.patch